Update protocol version from 8 to 9 across multiple test files and server implementations
- Updated the protocol version in various test files including `tst_DeviceIdentityStore.cpp`, `tst_ProjectManager.cpp`, `tst_RemoteSceneControllerLifecycle.cpp`, and others to ensure consistency with the new protocol version.
- Adjusted the `renderSchemaVersion` from 3 to 4 in relevant JSON objects to reflect schema updates.
- Modified the server-side protocol handling in files such as `device_auth.js`, `scene_timeline_validation.js`, and `server.js` to accommodate the new protocol version.
- Ensured all related tests and assertions are aligned with the updated protocol version to maintain functionality and integrity of the system.

+++ client/src/backend/domain/canvas/CanvasDocument.h
@@ -60,8 +60,8 @@ public:
     bool setPrimarySelectedMedia(const QString& mediaId);
     const SceneTimeline::SceneSettings& timelineSettings() const { return m_timelineSettings; }
     bool setTimelineSettings(const SceneTimeline::SceneSettings& settings);
-    qint64 timelinePositionMs() const { return m_timelinePositionMs; }
-    void setTimelinePosition(qint64 timeMs);
+    qreal timelinePositionMs() const { return m_timelinePositionMs; }
+    void setTimelinePosition(qreal timeMs);
     void evaluateTimeline();
     void select(const QString& mediaId, bool additive = false);
     void clearSelection();
@@ -160,7 +160,7 @@ private:
     QString m_primarySelectedMediaId;
     QStringList m_selectionActivationOrder;
     SceneTimeline::SceneSettings m_timelineSettings;
-    qint64 m_timelinePositionMs = 0;
+    qreal m_timelinePositionMs = 0;
     bool m_evaluatingTimeline = false;
     QList<ScreenInfo> m_screens;
     QHash<int, QRectF> m_screenRects;
